// src/utils/models/vecinos/editVecino.ts
export class EditarVecino{
    email: string;
    neighbor: neighbor;
    num_inmueble: string;
    }

export class neighbor{
    _id: string;
    id_usuario: string;
    num_inmueble: string;
    id_establecimiento: string;
    email: string;
    estado: string;
    nombres: string;
    numero_celular: string;
    apellidos: string;
    tipo_doc : string;
    identificacion: string;
    sexo: string;
    fecha_nacimiento: string;
    foto: string;
    }